# backend/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "economae API",
    "version": "1.0.0",
    "description": "Contrato entre o app Android e o backend. Gerado de backend/src/api/v1/definicoes.ts por `npm run contrato`; o teste falha se este arquivo divergir do código. Dinheiro sempre em centavos inteiros. Toda rota exige o cabeçalho X-Tenant; rotas com sessão exigem Authorization: Bearer."
  },
  "security": [
    {
      "sessao": []
    }
  ],
  "paths": {
...
    "/v1/auth/email/codigo": {
      "post": {
        "operationId": "pedirCodigoEmail",
        "summary": "Envia um código de 6 dígitos para o e-mail",
        "description": "Responde igual para e-mail cadastrado ou não. Só o código mais recente vale. Limites por e-mail e por IP, por hora.",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/Tenant"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  }
                },
                "required": [
                  "email"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Código enviado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "validadeMinutos": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "validadeMinutos"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida. Códigos: TENANT_INVALIDO, REQUISICAO_INVALIDA.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Limite de tentativas; ver Retry-After. Códigos: LIMITE_EXCEDIDO.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Dependência indisponível. Códigos: ENVIO_FALHOU.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/email/verificar": {
      "post": {
        "operationId": "verificarCodigoEmail",
        "summary": "Troca e-mail e código por uma sessão",
        "description": "Cria a conta no primeiro acesso. Cada código aceita um número limitado de tentativas.",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/Tenant"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  },
                  "codigo": {
                    "type": "string",
                    "pattern": "^[0-9]{6}$"
                  }
                },
                "required": [
                  "email",
                  "codigo"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sessão criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sessao"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida. Códigos: TENANT_INVALIDO, REQUISICAO_INVALIDA.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem sessão válida. Códigos: CODIGO_INVALIDO.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Limite de tentativas; ver Retry-After. Códigos: TENTATIVAS_ESGOTADAS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/google": {
      "post": {
        "operationId": "entrarComGoogle",
        "summary": "Troca um ID token do Google por uma sessão",
        "description": "Aceita só e-mail verificado pelo Google. Se já existe conta com o mesmo e-mail, vincula o Google a ela.",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/Tenant"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idToken": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 4096
                  }
                },
                "required": [
                  "idToken"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sessão criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sessao"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida. Códigos: TENANT_INVALIDO, REQUISICAO_INVALIDA.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem sessão válida. Códigos: TOKEN_GOOGLE_INVALIDO.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Proibido. Códigos: LOGIN_GOOGLE_INDISPONIVEL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Conflito. Códigos: EMAIL_VINCULADO_A_OUTRO_GOOGLE.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Dependência indisponível. Códigos: GOOGLE_INDISPONIVEL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/eu": {
      "get": {
        "operationId": "obterUsuario",
        "summary": "Conta, plano, termos e consentimento",
        "parameters": [
          {
            "$ref": "#/components/parameters/Tenant"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida. Códigos: TENANT_INVALIDO.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem sessão válida. Códigos: SESSAO_INVALIDA.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "excluirConta",
        "summary": "Exclui a conta e todos os dados pessoais",
        "description": "Apaga preferências, sessões, entregas, assinaturas e códigos de login. Não tem volta.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Tenant"
          }
        ],
        "responses": {
          "204": {
            "description": "Conta excluída"
          },
          "400": {
            "description": "Requisição inválida. Códigos: TENANT_INVALIDO.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem sessão válida. Códigos: SESSAO_INVALIDA.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/v1/eu/termos": {
      "put": {
        "operationId": "aceitarTermos",
        "summary": "Registra o aceite da versão vigente dos termos",
        "description": "O app envia a versão que mostrou; se ela não é mais a vigente, recebe 409 e mostra a nova.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Tenant"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "versao": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "versao"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida. Códigos: TENANT_INVALIDO, REQUISICAO_INVALIDA.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem sessão válida. Códigos: SESSAO_INVALIDA.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Conflito. Códigos: TERMOS_DESATUALIZADOS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/v1/eu/notificacoes": {
      "put": {
        "operationId": "definirConsentimentoNotificacoes",
        "summary": "Dá ou retira o consentimento para notificações",
        "description": "Separado do aceite dos termos (LGPD). Sem consentimento, nenhum push é enviado.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Tenant"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "consentidas": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "consentidas"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida. Códigos: TENANT_INVALIDO, REQUISICAO_INVALIDA.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem sessão válida. Códigos: SESSAO_INVALIDA.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessao": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token opaco devolvido pelo login."
      }
    },
    "parameters": {
      "Tenant": {
        "name": "X-Tenant",
        "in": "header",
        "required": true,
        "description": "Slug do tenant, fixo no build do app.",
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9-]+$"
        }
      }
    },
    "schemas": {
...
      "Erro": {
        "type": "object",
        "properties": {
          "erro": {
            "type": "object",
            "properties": {
              "codigo": {
                "type": "string",
                "description": "Estável; o app decide por ele."
              },
              "mensagem": {
                "type": "string"
              }
            },
            "required": [
              "codigo",
              "mensagem"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "erro"
        ],
        "additionalProperties": false
      },
...
      "Sessao": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Enviar em Authorization: Bearer <token>."
          },
          "expiraEm": {
            "type": "string",
            "format": "date-time",
            "description": "Renovada a cada uso; expira após o período sem uso."
          },
          "novoUsuario": {
            "type": "boolean"
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          }
        },
        "required": [
          "token",
          "expiraEm",
          "novoUsuario",
          "usuario"
        ],
        "additionalProperties": false
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "plano": {
            "type": "string",
            "enum": [
              "gratuito",
              "premium"
            ]
          },
          "limites": {
            "type": "object",
            "properties": {
              "maxCeps": {
                "type": "integer"
              }
            },
            "required": [
              "maxCeps"
            ],
            "additionalProperties": false
          },
          "termos": {
            "type": "object",
            "properties": {
              "versaoVigente": {
                "type": "string"
              },
              "versaoAceita": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "aceitosEm": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "pendente": {
                "type": "boolean",
                "description": "true enquanto a versão aceita não for a vigente."
              }
            },
            "required": [
              "versaoVigente",
              "versaoAceita",
              "aceitosEm",
              "pendente"
            ],
            "additionalProperties": false
          },
          "notificacoes": {
            "type": "object",
            "properties": {
              "consentidas": {
                "type": "boolean"
              },
              "consentidasEm": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              }
            },
            "required": [
              "consentidas",
              "consentidasEm"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "id",
          "email",
          "plano",
          "limites",
          "termos",
          "notificacoes"
        ],
        "additionalProperties": false
      }
    }
  }
}